Refactors most tests, some classes
