--- lib/playground.rb
@@ -1,7 +1,0 @@
-#iteration 4
-# sales_analyst = sales_engine.analyst
-# Find out the total revenue for a given date:
-#
-require './lib/invoice_item'
-require './test/test_helper'
-require './lib/sales_analyst'
