+++ test/sales_engine_test.rb
@@ -10,24 +10,24 @@ require './lib/customer'
 require './lib/customer_repository'
 
 class SalesEngineTest < Minitest::Test
-  # def setup
-  #   @se = SalesEngine.from_csv({
-  #     :items     => "./data/items.csv",
-  #     :merchants => "./data/merchants.csv",
-  #     :invoices => "./data/invoices.csv",
-  #     :invoice_items => "./data/invoice_items.csv",
-  #     :transactions => "./data/transactions.csv",
-  #     :customers => "./data/customers.csv"
-  #   })
-  # end
+  def setup
+    @se = SalesEngine.from_csv({
+      :items     => "./data/items.csv",
+      :merchants => "./data/merchants.csv",
+      :invoices => "./data/invoices.csv",
+      :invoice_items => "./data/invoice_items.csv",
+      :transactions => "./data/transactions.csv",
+      :customers => "./data/customers.csv"
+    })
+  end
 
   def test_it_exists
-    skip
+    # skip
     assert_instance_of SalesEngine, @se
   end
 
   def test_it_can_load_csv_files
-    skip
+    # skip
     assert_instance_of MerchantRepository, @se.merchants
     assert_instance_of ItemRepository, @se.items
     assert_instance_of InvoiceRepository, @se.invoices
@@ -37,7 +37,7 @@ class SalesEngineTest < Minitest::Test
   end
 
   def test_it_can_load_files_correctly
-    skip
+    # skip
     merchant_2 = Merchant.new({id: 12334112, name: "Candisart"})
     assert_equal merchant_2.id, @se.merchants.all[1].id
     assert_equal merchant_2.name, @se.merchants.all[1].name
@@ -67,7 +67,7 @@ class SalesEngineTest < Minitest::Test
   end
 
   def test_it_can_create_analyst
-    skip
+    # skip
     sa = @se.analyst
     assert_instance_of SalesAnalyst, sa
   end
