+++ test/test_helper.rb
@@ -3,5 +3,3 @@ SimpleCov.start
 require 'minitest/autorun'
 require 'minitest/pride'
 require 'pry'
-
-# require './test/test_helper'  <--- add this to every teswt
