+++ test/transaction_test.rb
@@ -16,6 +16,4 @@ class TransactionTest < Minitest::Test
     assert_instance_of Transaction, @tr
   end
 
-  
-
 end
